Fix: Missing Parameters Preventing to Checkout with an Address that Failed Verification
Fix: The function wrapping for the integration with FireCheckout wasn't passing the required parameters for the skin/frontend/base/default/tm/firecheckout/js/firecheckout.js::FireCheckout.prototype.save() function which prevented customers from checking out with an address that didn't passed the verification. To proceed and place an order even when the address isn't validated, a URL parameter is passed as the urlSuffix parameter of the save function.
Note: The address verification has to be tested with an address located within the United States.